🚚 Move definition of semverDecrement function to test/helpers/index.ts
